% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_batch.R
\name{assess_batch}
\alias{assess_batch}
\title{Assess Leaf Area for a Folder of Images}
\usage{
assess_batch(
  input_dir,
  workers = parallel::detectCores(all.tests = T) - 1,
  threshold = 120,
  cut_off = 10000,
  output_dir = NULL,
  combine = FALSE,
  res = 300
)
}
\arguments{
\item{input_dir}{The path to a folder (directory) that contains images, for which you want to assess the leaf area.}

\item{workers}{How many cores to use? Default is to use all available minus one. Only relevant when assessing a folder, ignored otherwise. Parallelization occurs within Python.}

\item{threshold}{A value between 0 (black) and 255 (white) for classification of background and leaf pixels. Default = 120}

\item{cut_off}{Clusters with fewer pixels than this value will be discarded. Default is 10000, which is about 3.3 pixels x 3.3 pixels, in a 300 dots per inch image.}

\item{output_dir}{The directory where to save the processed images. Respects tilde expansion. By default, processed images are not saved. This must not be a directory that already exists, so that existing files are not over-written.}

\item{combine}{If true the total area will be returned; otherwise each segment will be returned separately}

\item{res}{Image resolution, in dots per inch (DPI); if False the resolution will be read from the exif tag. Default is 300 DPI.}
}
\value{
The assessed leaf areas for an entire folder of images, in cm^2, is returned as a \code{data.frame}.
}
\description{
This function assesses the area of leaves taken from images. Ideally, the images will be obtained from a flatbed scanner, and will include resolution information in the exif metadata. Images derived from cameras can also be used, but in that case, you will probably need to supply the resolution explictly, using the argument res.
}
\examples{
input_dir <- area_example("prepared")
output_dir <- area_example("processed")

# If the argument output_dir is omitted,  do not save  processed images
assess_batch(input_dir, res = 400, combine = FALSE)
\dontrun{
assess_batch(input_dir, output_dir = "processed2", cut_off = 10000, res = 400)
}

}
