% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_batch.R
\name{preprocess_batch}
\alias{preprocess_batch}
\title{Prepare a Folder of Images for the Assessment of Leaf Area}
\usage{
preprocess_batch(
  input_dir,
  output_dir,
  workers = parallel::detectCores(all.tests = T) - 1,
  crop = 0,
  red_scale = 0,
  mask_scale = 0,
  mask_offset_x = 0,
  mask_offset_y = 0
)
}
\arguments{
\item{input_dir}{The path to a folder (directory) that contains images, that need to be prepared for assessment.}

\item{output_dir}{The directory where to save the processed images, so that they can be assessed using \code{assess}. Respects tilde expansion. This must not be a directory that already exists, so that existing files are not over-written.}

\item{workers}{How many cores to use? Default is to use all available minus one. Only relevant when processing a folder, ignored otherwise. Parallelization occurs within Python.}

\item{crop}{Should the margins of the image be cropped? Cropping occurs before the other operations, meaning that they are performed on the cropped image. Value is one integer; number of pixels to be removed from (top, bottom, left, right) of the image. Default = 0.}

\item{red_scale}{How many pixels wide should the side of the scale should be? Default = 0.}

\item{mask_scale}{How many pixels should each side of the masking window be? Default = 0.}

\item{mask_offset_x}{Offset for positioning the masking window in number of pixels from right to left of the image.}

\item{mask_offset_y}{for positioning the masking window in number of pixels from bottom to top of the image.}
}
\value{
No value is returned. The side effect is that processed images are saved in the directory specified in \code{output_dir}.
}
\description{
\code{preprocess_batch} prepares images of leaves for assessment using \code{assess}. Not all images require preprocessing. It provides options to crop away coloured margins from an image, to mask out an un-needed scale bar, or to insert a scale bar of a particular size. Either a single image or single directory of images can be processed at once. If the latter, processing can be performed in parallel.
}
\examples{
input_dir <- area_example("raw")
output_dir <- area_example("prepared")

#Demonstrate batch processing
\dontrun{
preprocess_batch(input_dir, output_dir, crop = 30)
preprocess_batch(input_dir, output_dir, workers = 5, crop = 30, mask_scale = 500, mask_offset_x = 0, mask_offset_y = 2600)
}
}
